Move WarningHandlingLevel to presto-common
