+++ presto-common/src/main/java/com/facebook/presto/common/WarningHandlingLevel.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.execution.warnings;
+package com.facebook.presto.common;
 
 public enum WarningHandlingLevel
 {
