+++ presto-main/src/main/java/com/facebook/presto/execution/warnings/WarningCollectorFactory.java
@@ -13,6 +13,7 @@
  */
 package com.facebook.presto.execution.warnings;
 
+import com.facebook.presto.common.WarningHandlingLevel;
 import com.facebook.presto.spi.WarningCollector;
 
 public interface WarningCollectorFactory
