+++ presto-main/src/test/java/com/facebook/presto/execution/TestPartialResultQueryManager.java
@@ -19,7 +19,7 @@ import com.facebook.presto.execution.warnings.WarningCollectorConfig;
 import com.facebook.presto.spi.WarningCollector;
 import org.testng.annotations.Test;
 
-import static com.facebook.presto.execution.warnings.WarningHandlingLevel.NORMAL;
+import static com.facebook.presto.common.WarningHandlingLevel.NORMAL;
 import static org.testng.Assert.assertEquals;
 
 public class TestPartialResultQueryManager
